refactor: move color writer implementation to logging package and update usage in server setup
